print first 50 Fibonacci numbers

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,7 +7,7 @@
  */
 int main(void)
 {
-	int i, num1, num2;
+	long int i, num1, num2;
 
 	num1 = 1;
 	num2 = 2;
